Move firebase functions to separate file
